Borrow existing function to produce nice end of input
In case of failure, the default `endOfInput` produces not enough
information to figure out where exactly the parsing errored. Use the
`niceEndOfinput` from Purebred to produce a better message so debugging
is easier.

Note, this is slightly changed, because sometimes just knowing that the
error happened at a whitespace char at offset X doesn't help much. Also
show the remaining input so it's easier to figure out where exactly the
error is.